+++ src/components/Buttons/PagenationButtons.tsx
@@ -1,10 +1,7 @@
 import { useState } from 'react';
-import Button from '../../Layout/Button';
-import ArrowForwardIcon from '@/assets/images/arrow-forward.svg';
-import ArrowBackwardIcon from '@/assets/images/arrow-backward.svg';
 
 export default function PagenationButtons() {
-  const [allPage, setAllPage] = useState(1);
+  const [allPage, setAllPage] = useState(10);
   const [nowPage, setNowPage] = useState(1); // Page 값들은 임시로 만들어 두었습니다. 추후에 api 연동해서 페이지네이션 구현 시 설정해줘야 합니다.
 
   const handleBackwardButtonClick = () => {
@@ -27,20 +24,27 @@ export default function PagenationButtons() {
         {allPage} 페이지 중 {nowPage}
       </span>
       <div className="flex items-center">
-        <Button
+        <button
           className="flex items-center justify-center w-[4rem] h-[4rem] bg-white border rounded-l-[0.4rem] border-gray_D9D9D9"
           type="button"
           onClick={handleBackwardButtonClick}
         >
-          <img src={ArrowBackwardIcon} alt="뒤로 가기 아이콘" />
-        </Button>
-        <Button
+          <img
+            src="../../assets/images/arrow-backward.svg"
+            alt="뒤로 가기 아이콘"
+          />
+        </button>
+        <button
           className="flex items-center justify-center w-[4rem] h-[4rem] bg-white border rounded-r-[0.4rem] border-gray_D9D9D9"
           type="button"
           onClick={handleForwardButtonClick}
         >
-          <img className="" src={ArrowForwardIcon} alt="앞으로 가기 아이콘" />
-        </Button>
+          <img
+            className=""
+            src="@/assets/images/arrow-forward.svg"
+            alt="앞으로 가기 아이콘"
+          />
+        </button>
       </div>
     </div>
   );
