--- src/components/ColorSelector/DashboardColorSelector.tsx
@@ -1,7 +1,0 @@
-import ColorSelector from '../Layout/ColorSelector';
-
-export default function DashboardColorSelector() {
-  return (
-    <ColorSelector colors={['green', 'purple', 'orange', 'blue', 'pink']} />
-  );
-}
