+++ src/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App';
+import App from './App.tsx';
 import './styles/global.css';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
