+++ src/types/common.ts
@@ -1,5 +1,1 @@
-export interface StringMap {
-  [key: string]: string;
-}
-
 export type ButtonType = JSX.IntrinsicElements['button']['type'];
